Guardrails/pyproject.toml
[project]
name = "guardrails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.3",
    "guardrails-ai==0.6.0",
    "guardrails-api-client>=0.4.0a1",
    "openai>=1.56.2",
]
